score: Fix undefined behaviour
Cast POSIX error codes to unsigned int to avoid undefined behaviour in
case of PTHREAD_BARRIER_SERIAL_THREAD which is -1.

+++ cpukit/score/include/rtems/score/status.h
@@ -53,7 +53,7 @@ typedef enum {
  */
 #if defined(RTEMS_POSIX_API)
   #define STATUS_BUILD( classic_status, posix_status ) \
-    ( ( ( posix_status ) << 8 ) | ( classic_status ) )
+    ( ( ( (unsigned int) ( posix_status ) ) << 8 ) | ( classic_status ) )
 #else
   #define STATUS_BUILD( classic_status, posix_status ) \
     ( classic_status )
